DOC-7961 -- Suppress http headers https://issues.couchbase.com/browse/DOC-7961
Suppress headers in Sync Gateway HTTP responses.
As a developer, I should be able to customize the Sync Gateway response to the root path; typically so as not to reveal the version of the Sync Gateway to HTTP requests to the root path.

(cherry picked from commit 5ac8cb335bd3d6f4bf8842ef52b384134a165997)